pyproject.toml
[tool.poetry]
name = "ko-sayalab-deep"
version = "0.1.0"
description = ""
authors = ["Haruki Inoue <haruki@hacarus.com>"]
packages = [
  {include = "*", from = "modules"},
]

[tool.poetry.dependencies]
python = "3.8.6"
numpy = "^1.23.2"
pandas = "^1.4.3"
matplotlib = "^3.5.3"
scipy = "^1.9.0"
scikit-learn = "^1.1.2"
scikit-image = "^0.19.3"
opencv-python = "^4.6.0"
seaborn = "^0.11.2"
torch = "^1.12.1"
openslide-python = "^1.2.0"
toml = "^0.10.2"
tqdm = "^4.64.0"
xlrd = "^2.0.1"
openpyxl = "^3.0.10"
hydra-core = "^1.2.0"
cupy = "^11.0.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.4.5"
flake8 = "^5.0.4"
black = "^22.6.0"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pylint-gitlab = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
